requirements.txt
absl-py==1.0.0
amqp==5.1.0
arabic-reshaper==2.1.3
asgiref==3.4.1
astunparse==1.6.3
autopep8==1.6.0
billiard==3.6.4.0
bleach==4.1.0
cachetools==4.2.4
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
convertdate==2.3.2
cycler==0.11.0
Cython==0.29.26
dj-database-url==0.5.0
Django==3.2.10
django-crispy-forms==1.7.2
django-on-heroku==1.1.2
django-seed==0.3.1
django-storages==1.12.3
django-tinymce==3.4.0
djangorestframework==3.13.1
docutils==0.18.1
ephem==4.1.3
export==0.1.2
Faker==13.3.4
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.4
fsspec==2022.3.0
future==0.18.2
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.43.0
gunicorn==20.1.0
h5py==3.6.0
hijri-converter==2.2.2
html5lib==1.1
idna==3.3
image==1.5.33
imbalanced-learn==0.8.1
imblearn==0.0
import-export==0.2.67.dev6
importlib-metadata==4.10.0
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==23.5.0
kiwisolver==1.3.2
kombu==5.2.4
korean-lunar-calendar==0.2.1
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplot==0.1.9
matplotlib==3.5.1
mysql==0.0.3
mysql-connector-python==8.0.28
mysqlclient==2.1.0
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
passlib==1.7.4
Pillow==9.0.1
pkginfo==1.8.2
prompt-toolkit==3.0.28
protobuf==3.19.4
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
Pygments==2.11.1
pyloco==0.0.139
PyMeeus==0.5.11
pyparsing==3.0.6
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.5
python-dotenv==0.20.0
pytz==2019.1
pywin32-ctypes==0.2.0
readme-renderer==32.0
reportlab==3.6.5
requests==2.27.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
rsa==4.8
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-git==1.2
SimpleWebSocketServer==0.1.1
six==1.13.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.2
toposort==1.7
tqdm==4.62.3
tweepy==4.5.0
twine==3.7.1
typing==3.7.4.3
typing_extensions==4.0.1
tzdata==2021.5
Unipath==1.1
urllib3==1.26.7
ushlex==0.99.1
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==2.0.3
whitenoise==5.3.0
wrapt==1.13.3
WTForms==3.0.1
xhtml2pdf==0.2.5
zipp==3.7.0
